dev(workspace): use publishable libs from workspace
In order to avoid the complexity related to yarn/npm link and unlink
of publishable libs built as npm package, we add path mappings to
the tsconfig to pull the imports out of the repo source files.
Publishable libraries used as peer dependencies in other libraries will
still need to be installed from the npm repo as normal dependencies
in package.json. Independently of that, imports in the workspace will
still be taken from the source modules and will be built as part of the
apps without any restriction on the use of Ivy or not.

https://github.com/angular/vscode-ng-language-service/issues/665
https://github.com/angular/vscode-ng-language-service/issues/645
https://github.com/angular/vscode-ng-language-service/issues/827