feat(ngx-highlight): decouple debounce time from style config
BREAKING CHANGE: in normal use of the directive, it is much less likely
to reset the debounce time then it is to change the highlight style.
Therefore, instead of treating the debounce time as part of the style
config, it is more appropriate to treat it separately.

BREAKING CHANGE: default values for style config and for debounce
time are now constants exported by the module in the public api. This
is cleaner.